requirements.txt
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
psycopg2-binary==2.9.1
python-dotenv==0.15.0
